health: #184 reorganized criteria tests a bit
