+++ packages/criteria/src/test/reduction/and-criteria.test.ts
@@ -1,4 +1,4 @@
-import { reducing } from "./reducing.fn";
+import { reducing } from "../tools/reducing.fn";
 
 describe("reducing: and-criteria", () => {
     describe("full reduction", () => {
