+++ packages/criteria/src/test/reduction/binary.test.ts
@@ -1,4 +1,4 @@
-import { reducing } from "./reducing.fn";
+import { reducing } from "../tools/reducing.fn";
 
 describe("reducing: binary", () => {
     describe("full reduction", () => {
