+++ packages/criteria/src/test/reduction/in-range.test.ts
@@ -1,4 +1,4 @@
-import { reducing } from "./reducing.fn";
+import { reducing } from "../tools/reducing.fn";
 
 describe("reducing: in-range", () => {
     describe("full reduction", () => {
