+++ packages/criteria/src/test/tools/inverting.fn.ts
@@ -1,5 +1,5 @@
-import { Criterion } from "../lib/criterion/criterion";
-import { parseCriteria } from "../lib/parser/parse-criteria.fn";
+import { Criterion } from "../../lib/criterion/criterion";
+import { parseCriteria } from "../../lib/parser/parse-criteria.fn";
 
 function parse<T extends Criterion | string>(item: T): Criterion {
     if (typeof item === "string") {
