+++ packages/criteria/src/test/tools/reducing.fn.ts
